% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_binary.R
\name{code.binary}
\alias{code.binary}
\title{Code one binary variable yes / no}
\usage{
code.binary(dat, varname, labs = c("No", "Yes"))
}
\arguments{
\item{dat}{the data you are using}

\item{varname}{the variable you want to recode}

\item{labs}{the label of the binary variable. defaults to yes/no}
}
\description{
Code one binary variable yes / no
}
