% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssicToXLSX.R
\name{mssicToXLSX}
\alias{mssicToXLSX}
\title{Function to read in MSSIC data, recode and converto xlsx}
\usage{
mssicToXLSX(date, fp, all = F, levels = F)
}
\arguments{
\item{date}{date in \%m-\%d-\%Y format}

\item{fp}{filepath to save the sas files}

\item{all}{whether or not it should return all 3 (uncoded) datasets or just the merged dataset, default just the merged}
}
\description{
e
}
