% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notone.R
\name{notone}
\alias{notone}
\title{A function to make logical equivalencies with SAS}
\usage{
notone(x, y = 1)
}
\arguments{
\item{x}{input number}

\item{y}{defaults to 1 hence the function name but can be any number.}
}
\description{
Makes the equivalent for ne in SAS code as R != returns NA if the value is NA but we want 2 != NA to be TRUE not NA
}
\examples{
NA == 1
notone(NA)
}
\keyword{mssic}
