% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mssic_data.R
\name{read.mssic.data}
\alias{read.mssic.data}
\title{Function to read in MSSIC data just by giving a dat}
\usage{
read.mssic.data(date)
}
\arguments{
\item{date}{date in \%m-\%d-\%Y format}
}
\description{
e
}
