% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.mssic}
\alias{recode.mssic}
\title{Main ?package function to recode data}
\usage{
recode.mssic(dat, levels = T)
}
\arguments{
\item{dat}{merged mssic dataset}

\item{levels}{logical value asking whether number of levels should be computed. Defaults to T as usually they are wanted, but sometimes takes a short while to run hence the option}
}
\description{
Recodes a merged MSSIC dataset according to the standards set out by the MSSIC analytics team
All values not yet verified to match SAS and STATA.
}
