% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_sites.R
\name{remove.sites}
\alias{remove.sites}
\title{Remove records from disabled sites}
\usage{
remove.sites(dat, additional.sites = NULL)
}
\arguments{
\item{additional.sites}{sites not listed as of 06/2024. Defaults to NULL}
}
\description{
Remove records from disabled sites
}
