% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add.labels.after.recode}
\alias{add.labels.after.recode}
\title{Label recoded MSSIC data for documentation and visualization}
\usage{
add.labels.after.recode(dat, missing.as.factor = T)
}
\arguments{
\item{dat}{recoded MSSIC dataset}

\item{missing.as.factor}{defaults to T, whether to add missing as a factor or leave it as NA}
}
\description{
Label recoded MSSIC data for documentation and visualization
}
